Moved RandomState into Random namespace
