+++ Src/Newtonsoft.Json.UnityConverters.Tests/ConvertingUnityTypes/Random/RandomStateTests.cs
@@ -4,30 +4,30 @@ using System.Linq;
 using System.Reflection;
 using NUnit.Framework;
 using UnityEngine;
-using Random = UnityEngine.Random;
+using State = UnityEngine.Random.State;
 
-namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes
+namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes.Random
 {
-    public class RandomStateTests : ValueTypeTester<Random.State>
+    public class RandomStateTests : ValueTypeTester<State>
     {
-        private static readonly FieldInfo[] _randomStateFields = typeof(Random.State).GetFields(BindingFlags.NonPublic | BindingFlags.Instance)
+        private static readonly FieldInfo[] _randomStateFields = typeof(State).GetFields(BindingFlags.NonPublic | BindingFlags.Instance)
             .OrderBy(o => o.Name)
             .WhereNotNullRef()
             .ToArray();
 
-        public static readonly IReadOnlyCollection<(Random.State deserialized, object anonymous)> representations = new (Random.State, object)[] {
-            (new Random.State(), new { s0 = 0, s1 = 0, s2 = 0, s3 = 0 }),
+        public static readonly IReadOnlyCollection<(State deserialized, object anonymous)> representations = new (State, object)[] {
+            (new State(), new { s0 = 0, s1 = 0, s2 = 0, s3 = 0 }),
             (CreateState(1,2,3,4), new { s0 = 1, s1 = 2, s2 = 3, s3 = 4 }),
         };
 
-        protected override bool AreEqual(Random.State a, Random.State b)
+        protected override bool AreEqual(State a, State b)
         {
             if (_randomStateFields.Length != 4)
             {
-                throw new InvalidOperationException("Was unable to find all four random state fields from the UnityEngine.Random.State type.");
+                throw new InvalidOperationException("Was unable to find all four random state fields from the UnityEngine.State type.");
             }
 
-            foreach(FieldInfo field in _randomStateFields)
+            foreach (FieldInfo field in _randomStateFields)
             {
                 if (!Equals(field.GetValue(a), field.GetValue(b)))
                 {
@@ -38,25 +38,25 @@ namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes
             return true;
         }
 
-        protected override string ToString(Random.State value)
+        protected override string ToString(State value)
         {
             if (_randomStateFields.Length != 4)
             {
-                throw new InvalidOperationException("Was unable to find all four random state fields from the UnityEngine.Random.State type.");
+                throw new InvalidOperationException("Was unable to find all four random state fields from the UnityEngine.State type.");
             }
 
             return $"[{string.Join(", ", _randomStateFields.Select(o => o.GetValue(value)))}]";
         }
 
-        private static Random.State CreateState(int s0, int s1, int s2, int s3)
+        private static State CreateState(int s0, int s1, int s2, int s3)
         {
             if (_randomStateFields.Length != 4)
             {
-                throw new InvalidOperationException("Was unable to find all four random state fields from the UnityEngine.Random.State type.");
+                throw new InvalidOperationException("Was unable to find all four random state fields from the UnityEngine.State type.");
             }
 
             string json = $@"{{""s0"":{s0},""s1"":{s1},""s2"":{s2},""s3"":{s3}}}";
-            return JsonUtility.FromJson<Random.State>(json);
+            return JsonUtility.FromJson<State>(json);
         }
 
         [Test]
@@ -65,7 +65,7 @@ namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes
             // Arrange
             FieldInfo field = _randomStateFields[0];
 
-            object boxed = new Random.State();
+            object boxed = new State();
             Assert.AreEqual(0, field.GetValue(boxed));
 
             int newValue = 5;
@@ -83,7 +83,7 @@ namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes
             // Arrange
             FieldInfo field = _randomStateFields[0];
 
-            var value = new Random.State();
+            var value = new State();
             Assert.AreEqual(0, field.GetValue(value));
 
             int newValue = 5;
